fixes bug that was causing certain containers to return bad usage percentages
